checks for NaNs during non-linear solving

+++ src/canareq/canareq.hpp
@@ -21,7 +21,7 @@ class canareq{
         void delete_2d_double_array(double **array);
     
         // subroutines
-        double thrust(int, float);
+        double thrust(int, double);
         double mat3(double**, double*);
         void linearModel(); // generate 'best' values
         void nonlinearModel();
@@ -47,28 +47,28 @@ class canareq{
         int inewton,imarg,itfd,nalmin,nalmax,n,incidence;
         //parameter(lxx=101,ndatx=21)
         int *kxtrm;
-        float pi,eps,degrad,radeg,us3,omega,epser,rho,amu,mass;
-        float xcg,statmarg,amlb,mg,bm,cxm,cam,am,xac,tfd,dClmdtf;
-        float dCmmdtf,dCdtf0,dCdtf1,dCdtf2,dm,em,tf,rf,hf,af;
-        float Cdbrake,aref,lref,hpower,Pcent,Ueq,Ref;
-        float T,dTdv,tr0,dtrdv,arm,prod,dum,dcz,dcxm,dcxp,incd;
-        float dCldam0,Clm00,dCmacdam0,dCmdam0,Cmacm00,Cmm00,Cdm,dClda;
-        float dCmda,aleq,aleqd,Cleq,dynaref,Cdeq,Cteq,Cweq,Clm0;
-        float beteq,Clmeq,Cmac0,Cmcg,dCldam,dCmacda,dCmdam;
-        float dCmeq,Cmac,amarg,dcxdcz,d2cxdcz2,Cdmeq,Cl0,Cm0,Cdm0,Cdim;
-        float daleq,dUeq,reyeq,reym,Cdf0,Cdfeq,dbeteq,beteqd,theq,theqd;
-        float ClCdeq,reseq,winglift,alphd,alph,Cl,Cm,Cd,Cl3,Cd2,ratio1;
-        float ratio2,Cmmeq,Cmeq,Cmm0,dyn,hpokwatt,lf,reyf,weq;
-        float reseq0,bc,cxc,cac,ac,xacc,dc,ec,tcd,tc,awc,acw,xacm;
-        float arc,dCmacdac0,Cmacc00,Clc00,dCldac0,Cmac00,Cmc00,dCmdac0;
-        float Clceq,canarlift,reyc,Cdic,Cdceq,Clc0,Cmacc0,Cmacm0,Cmacm;
-        float Cmc0,dCldac,dCmacdac,dCmacdam,dCmdac,Cdc0,Cmacc;
-        float aleq0,Ueq0,beteq0,Cleq0,Cdeq0,rcor,xcpm,xcpc,dCldalind;
-        float aleqB, UeqB, beteqB, CleqB, CdeqB, CmacB;
-        float zeng,rav,ruh,ar,reyr,Cdr0,Cdreq,dna,lna,reyn,an;
-        float Cdn0,Cdneq;
-        float *vr; float *tr;
-        float *cx; float *cz; float *cq; float *inc;
+        double pi,eps,degrad,radeg,us3,omega,epser,rho,amu,mass;
+        double xcg,statmarg,amlb,mg,bm,cxm,cam,am,xac,tfd,dClmdtf;
+        double dCmmdtf,dCdtf0,dCdtf1,dCdtf2,dm,em,tf,rf,hf,af;
+        double Cdbrake,aref,lref,hpower,Pcent,Ueq,Ref;
+        double T,dTdv,tr0,dtrdv,arm,prod,dum,dcz,dcxm,dcxp,incd;
+        double dCldam0,Clm00,dCmacdam0,dCmdam0,Cmacm00,Cmm00,Cdm,dClda;
+        double dCmda,aleq,aleqd,Cleq,dynaref,Cdeq,Cteq,Cweq,Clm0;
+        double beteq,Clmeq,Cmac0,Cmcg,dCldam,dCmacda,dCmdam;
+        double dCmeq,Cmac,amarg,dcxdcz,d2cxdcz2,Cdmeq,Cl0,Cm0,Cdm0,Cdim;
+        double daleq,dUeq,reyeq,reym,Cdf0,Cdfeq,dbeteq,beteqd,theq,theqd;
+        double ClCdeq,reseq,winglift,alphd,alph,Cl,Cm,Cd,Cl3,Cd2,ratio1;
+        double ratio2,Cmmeq,Cmeq,Cmm0,dyn,hpokwatt,lf,reyf,weq;
+        double reseq0,bc,cxc,cac,ac,xacc,dc,ec,tcd,tc,awc,acw,xacm;
+        double arc,dCmacdac0,Cmacc00,Clc00,dCldac0,Cmac00,Cmc00,dCmdac0;
+        double Clceq,canarlift,reyc,Cdic,Cdceq,Clc0,Cmacc0,Cmacm0,Cmacm;
+        double Cmc0,dCldac,dCmacdac,dCmacdam,dCmdac,Cdc0,Cmacc;
+        double aleq0,Ueq0,beteq0,Cleq0,Cdeq0,rcor,xcpm,xcpc,dCldalind;
+        double aleqB, UeqB, beteqB, CleqB, CdeqB, CmacB;
+        double zeng,rav,ruh,ar,reyr,Cdr0,Cdreq,dna,lna,reyn,an;
+        double Cdn0,Cdneq;
+        double *vr; double *tr;
+        double *cx; double *cz; double *cq; double *inc;
         double det,usdet,b1,b2,b3;
         double **aa; double *bb;
         //std::string title, prop;
